* Solved most of the problem from ticket #12, where the integrated flux was being calculated differently on different machines. Now casting the spatial size of a detection to a double.  * Solved ticket #13 as well, to allow compilation when PGPLOT is not available. Included moving cpgIsPS() to mycpgplot.cc.
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@321 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/Detection/detection.cc
@@ -363,7 +363,7 @@ void Detection::calcIntegFlux(float *fluxArray, long *dim, FitsHeader &head)
     int size = this->fluxUnits.size();
     std::string tailOfFluxUnits = this->fluxUnits.substr(size-5,size);
     if(tailOfFluxUnits == "/beam")
-      this->intFlux  *= this->getSpatialSize()/head.getBeamSize();
+      this->intFlux  *= double(this->getSpatialSize())/head.getBeamSize();
   }
 
 }
